+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Enhancements
+
+- Improve `Ethers.deploy/2` error handling
+
 ## v0.5.5 (2024-12-03)
 
 ### Enhancements